Move: 멤버 로직을 domain 디렉터리 하위로 이동
