+++ src/main/java/org/example/sivillage/SiVillageApplication.java
@@ -2,8 +2,10 @@ package org.example.sivillage;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing // auditing
 public class SiVillageApplication {
 
     public static void main(String[] args) {
