+++ src/main/java/org/example/sivillage/domain/brand/dto.java
@@ -1,0 +1,4 @@
+package org.example.sivillage.domain.brand;
+
+public class dto {
+}
