+++ src/main/java/org/example/sivillage/domain/member/vo/MemberRequest.java
@@ -1,0 +1,5 @@
+package org.example.sivillage.domain.member.vo;
+
+public class MemberRequest {
+}
+//test
