+++ src/main/java/org/example/sivillage/global/auth/CustomUserDetailsService.java
@@ -3,8 +3,8 @@ package org.example.sivillage.global.auth;
 import lombok.RequiredArgsConstructor;
 import org.example.sivillage.global.error.CustomException;
 import org.example.sivillage.global.error.ErrorCode;
-import org.example.sivillage.member.domain.Member;
-import org.example.sivillage.member.infrastructure.MemberRepository;
+import org.example.sivillage.domain.member.domain.Member;
+import org.example.sivillage.domain.member.infrastructure.MemberRepository;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
