+++ src/main/java/org/example/sivillage/global/error/ErrorCode.java
@@ -17,6 +17,7 @@ public enum ErrorCode {
      * 404 NOT_FOUND: 리소스를 찾을 수 없음
      */
     POSTS_NOT_FOUND(HttpStatus.NOT_FOUND, "게시글 정보를 찾을 수 없습니다."),
+    BRAND_NOT_FOUND(HttpStatus.NOT_FOUND, "해당하는 브랜드 명을 찾을 수 없습니다."),
 
     /*
      * 405 METHOD_NOT_ALLOWED: 허용되지 않은 Request Method 호출
@@ -43,6 +44,7 @@ public enum ErrorCode {
     ALREADY_FRIEND(HttpStatus.CONFLICT, "이미 친구로 추가된 회원입니다."),
     MEMBER_NOT_FOUND(HttpStatus.NOT_FOUND, "해당하는 회원을 찾을 수 없습니다."),
     UNAUTHORIZED_TOKEN(HttpStatus.UNAUTHORIZED, "권한 정보가 없는 토큰입니다."),
+
 
     // 권한 관련
     NO_AUTHORITY_FOUND(HttpStatus.NOT_FOUND, "회원의 권한 정보를 찾을 수 없습니다"),
