--- src/main/java/org/example/sivillage/member/vo/MemberRequestVo.java
@@ -1,5 +1,0 @@
-package org.example.sivillage.member.vo;
-
-public class MemberRequestVo {
-}
-//test
